refactor(multiple): various refactoring and updates
* expand README and DEVELOPER.md
* rename/reorganize files
* move grid/dis-related utilities to flopy
* move common test fixtures here from flopy
* move test fixture tests here from flopy
