--- modflow_devtools/__init__.py
@@ -1,9 +1,0 @@
-from .config import (
-    __author__,
-    __date__,
-    __description__,
-    __email__,
-    __maintainer__,
-    __status__,
-    __version__,
-)
