--- modflow_devtools/config.py
@@ -1,7 +1,0 @@
-__author__ = "Joseph D. Hughes"
-__date__ = "March 2, 2022"
-__version__ = "0.0.1"
-__maintainer__ = "Joseph D. Hughes"
-__email__ = "jdhughes@usgs.gov"
-__status__ = "Production"
-__description__ = """MODFLOW 6 developer tools."""
