+++ modflow_devtools/framework.py
@@ -1,773 +1,933 @@
 import os
 import shutil
-import sys
-import time
 
-import flopy
-import modflow_devtools as devtools
-import numpy as np
-
-
-class Framework:
-    def build(self, build_function, idx, exdir):
-        """
-        Build base and regression MODFLOW 6 models
-
-        Parameters
-        ----------
-        build_function : function
-            user defined function that builds a base model and optionally
-            builds a regression model. If a regression model is not built
-            then None must be returned from the function for the regression
-            model.
-        idx : int
-            counter that corresponds to exdir entry
-        exdir : str
-            path to regression model files
-        """
-        base, regression = build_function(idx, exdir)
-        base.write_simulation()
-        if regression is not None:
-            if isinstance(regression, flopy.mf6.MFSimulation):
-                regression.write_simulation()
-            else:
-                regression.write_input()
-
-    def run(self, config, workspace):
-        """
-        Run the MODFLOW 6 test and compare to existing head file or
-        MODFLOW-2005, MODFLOW-NWT, MODFLOW-USG, or MODFLOW-LGR run.
-
-        Parameters
-        ----------
-        config : TestConfig object
-            Modflow 6 test configuration object that runs the base and
-            regression models, compares the results, and tears down the
-            test if successful.
-        workspace : Pathlike
-            The test workspace
-        """
-        config.set_model(workspace, testModel=False)
-        config.run()
-        config.compare()
-        if config.exfunc is not None:
-            config.exfunc(config)
-
-
-class TestConfig:
-    # tell pytest this isn't a test class, don't collect it
-    __test__ = False
-
-    sfmt = "{:25s} - {}"
-    extdict = {
-        "hds": "head",
-        "hed": "head",
-        "bhd": "head",
-        "ucn": "concentration",
-        "cbc": "cell-by-cell",
-    }
-
-    def __init__(
-        self,
-        name,
-        targets,
-        exfunc=None,
-        exe_dict=None,
-        htol=None,
-        pdtol=None,
-        rclose=None,
-        idxsim=None,
-        cmp_verbose=True,
-        require_failure=None,
-        api_func=None,
-        mf6_regression=False,
-        make_comparison=True,
-    ):
-        self.targets = targets
-
-        for idx, arg in enumerate(sys.argv):
-            if arg[2:].lower() in list(self.targets.keys()):
-                key = arg[2:].lower()
-                exe0 = self.targets[key]
-                exe = os.path.join(os.path.dirname(exe0), sys.argv[idx + 1])
-                msg = (
-                    f"replacing {key} executable "
-                    + f'"{self.targets[key]}" with '
-                    + f'"{exe}".'
-                )
-                print(msg)
-                self.targets[key] = exe
-
-        if exe_dict is not None:
-            if not isinstance(exe_dict, dict):
-                msg = "exe_dict must be a dictionary"
-                assert False, msg
-            keys = list(self.targets.keys())
-            for key, value in exe_dict.items():
-                if key in keys:
-                    exe0 = self.targets[key]
-                    exe = os.path.join(os.path.dirname(exe0), value)
-                    msg = (
-                        f"replacing {key} executable "
-                        + f'"{self.targets[key]}" with '
-                        + f'"{exe}".'
-                    )
-                    print(msg)
-                    self.targets[key] = exe
-
-        msg = self.sfmt.format("Initializing test", name)
-        print(msg)
-        self.name = name
-        self.exfunc = exfunc
-        self.simpath = None
-        self.inpt = None
-        self.outp = None
-        self.coutp = None
-        self.api_func = api_func
-        self.mf6_regression = mf6_regression
-        self.make_comparison = make_comparison
-        self.action = None
-
-        # set htol for comparisons
-        if htol is None:
-            htol = 0.001
-        else:
-            msg = self.sfmt.format("User specified comparison htol", htol)
-            print(msg)
-
-        self.htol = htol
-
-        # set pdtol for comparisons
-        if pdtol is None:
-            pdtol = 0.001
-        else:
-            msg = self.sfmt.format(
-                "User specified percent difference comparison pdtol", pdtol
-            )
-            print(msg)
-
-        self.pdtol = pdtol
-
-        # set rclose for comparisons
-        if rclose is None:
-            rclose = 0.001
-        else:
-            msg = self.sfmt.format(
-                "User specified percent difference comparison rclose", rclose
-            )
-            print(msg)
-
-        self.rclose = rclose
-
-        # set index for multi-simulation comparisons
-        self.idxsim = idxsim
-
-        # set compare verbosity
-        self.cmp_verbose = cmp_verbose
-
-        # set allow failure
-        self.require_failure = require_failure
-
-        self.success = False
-
-    def __repr__(self):
-        return self.name
-
-    def set_model(self, pth, testModel=True):
-        """
-        Set paths to MODFLOW 6 model and associated comparison test
-        """
-        # make sure this is a valid path
-        if not os.path.isdir(pth):
-            assert False, f"{pth} is not a valid directory"
-
-        self.simpath = pth
-
-        # get MODFLOW 6 output file names
-        fpth = os.path.join(pth, "mfsim.nam")
-        mf6inp, mf6outp = devtools.get_mf6_files(fpth)
-        self.outp = mf6outp
-
-        # determine comparison model
-        self.setup_comparison(pth, pth, testModel=testModel)
-        # if self.mf6_regression:
-        #     self.action = "mf6-regression"
-        # else:
-        #     self.action = devtools.get_mf6_comparison(pth)
-        if self.action is not None:
-            if "mf6" in self.action or "mf6-regression" in self.action:
-                cinp, self.coutp = devtools.get_mf6_files(fpth)
-
-    def setup(self, src, dst):
-        msg = self.sfmt.format("Setup test", self.name)
-        self.originpath = src
-        self.simpath = dst
-        self.inpt, self.outp = devtools.setup_mf6(src=src, dst=dst)
-        time.sleep(0.5)
-        self.setup_comparison(src, dst)
-
-    def setup_comparison(self, src, dst, testModel=True):
-
-        # evaluate if comparison should be made
-        if not self.make_comparison:
-            return
-
-        # adjust htol if it is smaller than IMS outer_dvclose
-        dvclose = self._get_dvclose(dst)
-        if dvclose is not None:
-            dvclose *= 5.0
-            if self.htol < dvclose:
-                self.htol = dvclose
-
-        # get rclose to use with budget comparisons
-        rclose = self._get_rclose(dst)
-        if rclose is None:
-            rclose = 0.5
-        else:
-            rclose *= 5.0
-        self.rclose = rclose
-
-        # Copy comparison simulations if available
-        if self.mf6_regression:
-            action = "mf6-regression"
-            pth = os.path.join(dst, action)
-            if os.path.isdir(pth):
-                shutil.rmtree(pth)
-            shutil.copytree(dst, pth)
-        elif testModel:
-            action = devtools.setup_mf6_comparison(src, dst)
-        else:
-            action = devtools.get_mf6_comparison(dst)
-
-        self.action = action
-
-    def run(self):
-        """
-        Run the model and assert if the model terminated successfully
-        """
-        msg = self.sfmt.format("Run test", self.name)
-        print(msg)
-
-        # Set nam as namefile name without path
-        nam = None
-
-        # run mf6 models
-        exe = self.targets["mf6"]
-        msg = self.sfmt.format("using executable", exe)
-        print(msg)
-        try:
-            success, buff = flopy.run_model(
-                exe,
-                nam,
-                model_ws=self.simpath,
-                silent=False,
-                report=True,
-            )
-            msg = self.sfmt.format("MODFLOW 6 run", self.name)
-            if success:
-                print(msg)
-            else:
-                print(msg)
-        except:
-            msg = self.sfmt.format("MODFLOW 6 run", self.name)
-            print(msg)
-            success = False
-
-        # set failure based on success and require_failure setting
-        if self.require_failure is None:
-            msg = "MODFLOW 6 model did not terminate normally"
-            if success:
-                failure = False
-            else:
-                failure = True
-        else:
-            if self.require_failure:
-                msg = "MODFLOW 6 model should have failed"
-                if not success:
-                    failure = False
-                else:
-                    failure = True
-            else:
-                msg = "MODFLOW 6 model should not have failed"
-                if success:
-                    failure = False
-                else:
-                    failure = True
-
-        # print end of mfsim.lst to the screen
-        if failure:
-            fpth = os.path.join(self.simpath, "mfsim.lst")
-            msg = self._get_mfsim_listing(fpth) + msg
-
-        # test for failure
-        assert not failure, msg
-
-        self.nam_cmp = None
-        if success:
-            if self.action is not None:
-                if self.action.lower() == "compare":
-                    msg = self.sfmt.format("Comparison files", self.name)
-                    print(msg)
-                else:
-                    cpth = os.path.join(self.simpath, self.action)
-                    key = self.action.lower().replace(".cmp", "")
-                    exe = os.path.abspath(self.targets[key])
-                    msg = self.sfmt.format("comparison executable", exe)
-                    print(msg)
-                    if (
-                        "mf6" in key
-                        or "libmf6" in key
-                        or "mf6-regression" in key
-                    ):
-                        nam = None
-                    else:
-                        npth = devtools.get_namefiles(cpth)[0]
-                        nam = os.path.basename(npth)
-                    self.nam_cmp = nam
-                    try:
-                        if self.api_func is None:
-                            success_cmp, buff = flopy.run_model(
-                                exe,
-                                nam,
-                                model_ws=cpth,
-                                silent=False,
-                                report=True,
-                            )
-                        else:
-                            success_cmp, buff = self.api_func(
-                                exe, self.idxsim, model_ws=cpth
-                            )
-                        msg = self.sfmt.format(
-                            "Comparison run", self.name + "/" + key
-                        )
-                        print(msg)
-
-                        # print end of mfsim.lst to the screen
-                        if "mf6" in key:
-                            if not success:
-                                fpth = os.path.join(cpth, "mfsim.lst")
-                                print(self._get_mfsim_listing(fpth))
-
-                    except:
-                        success_cmp = False
-                        msg = self.sfmt.format(
-                            "Comparison run", self.name + "/" + key
-                        )
-                        print(msg)
-
-                    assert success_cmp, "Unsuccessful comparison run"
-
-    def compare(self):
-        """
-        Compare the model results
-
-        """
-        self.success = True
-
-        # evaluate if comparison should be made
-        if not self.make_comparison:
-            return
-
-        msgall = ""
-        msg = self.sfmt.format("Comparison test", self.name)
-        print(msg)
-
-        if self.action is not None:
-            cpth = os.path.join(self.simpath, self.action)
-            files_cmp = None
-            if self.action.lower() == "compare":
-                files_cmp = []
-                files = os.listdir(cpth)
-                for file in files:
-                    files_cmp.append(file)
-            elif "mf6" in self.action:
-                fpth = os.path.join(cpth, "mfsim.nam")
-                cinp, self.coutp = devtools.get_mf6_files(fpth)
-
-            head_extensions = (
-                "hds",
-                "hed",
-                "bhd",
-                "ahd",
-                "bin",
-            )
-            if "mf6-regression" in self.action:
-                success, msgall = self._compare_heads(
-                    msgall,
-                    extensions=head_extensions,
-                )
-                if not success:
-                    self.success = False
-            # non-regression runs - for new features
-            else:
-                files1 = []
-                files2 = []
-                exfiles = []
-                ipos = 0
-                for file1 in self.outp:
-                    ext = os.path.splitext(file1)[1][1:]
-
-                    if ext.lower() in head_extensions:
-
-                        # simulation file
-                        pth = os.path.join(self.simpath, file1)
-                        files1.append(pth)
-
-                        # look for an exclusion file
-                        pth = os.path.join(self.simpath, file1 + ".ex")
-                        if os.path.isfile(pth):
-                            exfiles.append(pth)
-                        else:
-                            exfiles.append(None)
-
-                        # Check to see if there is a corresponding compare file
-                        if files_cmp is not None:
-
-                            if file1 + ".cmp" in files_cmp:
-                                # compare file
-                                idx = files_cmp.index(file1 + ".cmp")
-                                pth = os.path.join(cpth, files_cmp[idx])
-                                files2.append(pth)
-                                txt = self.sfmt.format(
-                                    f"Comparison file {ipos + 1}",
-                                    os.path.basename(pth),
-                                )
-                                print(txt)
-                        else:
-                            if self.coutp is not None:
-                                for file2 in self.coutp:
-                                    ext = os.path.splitext(file2)[1][1:]
-
-                                    if ext.lower() in head_extensions:
-                                        # simulation file
-                                        pth = os.path.join(cpth, file2)
-                                        files2.append(pth)
-
-                            else:
-                                files2.append(None)
-
-                if self.nam_cmp is None:
-                    pth = None
-                else:
-                    pth = os.path.join(cpth, self.nam_cmp)
-
-                for ipos in range(len(files1)):
-                    file1 = files1[ipos]
-                    ext = os.path.splitext(file1)[1][1:].lower()
-                    outfile = os.path.splitext(os.path.basename(file1))[0]
-                    outfile = os.path.join(
-                        self.simpath, outfile + "." + ext + ".cmp.out"
-                    )
-                    if files2 is None:
-                        file2 = None
-                    else:
-                        file2 = files2[ipos]
-
-                    # set exfile
-                    exfile = None
-                    if file2 is None:
-                        if len(exfiles) > 0:
-                            exfile = exfiles[ipos]
-                            if exfile is not None:
-                                txt = self.sfmt.format(
-                                    f"Exclusion file {ipos + 1}",
-                                    os.path.basename(exfile),
-                                )
-                                print(txt)
-
-                    # make comparison
-                    success_tst = devtools.compare_heads(
-                        None,
-                        pth,
-                        precision="double",
-                        text=self.extdict[ext],
-                        outfile=outfile,
-                        files1=file1,
-                        files2=file2,
-                        htol=self.htol,
-                        difftol=True,
-                        # Change to true to have list of all nodes exceeding htol
-                        verbose=self.cmp_verbose,
-                        exfile=exfile,
-                    )
-                    msg = self.sfmt.format(
-                        f"{self.extdict[ext]} comparison {ipos + 1}",
-                        self.name,
-                    )
-                    print(msg)
-
-                    if not success_tst:
-                        self.success = False
-                        msgall += msg + " ... FAILED\n"
-
-            # compare concentrations
-            if "mf6-regression" in self.action:
-                success, msgall = self._compare_concentrations(msgall)
-                if not success:
-                    self.success = False
-
-            # compare cbc files
-            if "mf6-regression" in self.action:
-                cbc_extensions = (
-                    "cbc",
-                    "bud",
-                )
-                success, msgall = self._compare_budgets(
-                    msgall, extensions=cbc_extensions
-                )
-                if not success:
-                    self.success = False
-
-        assert self.success, msgall
-
-    def _get_mfsim_listing(self, lst_pth):
-        """Get the tail of the mfsim.lst listing file"""
-        msg = ""
-        ilen = 100
-        with open(lst_pth) as fp:
-            lines = fp.read().splitlines()
-        msg = "\n" + 79 * "-" + "\n"
-        if len(lines) > ilen:
-            i0 = -100
-        else:
-            i0 = 0
-        for line in lines[i0:]:
-            if len(line) > 0:
-                msg += f"{line}\n"
-        msg += 79 * "-" + "\n\n"
-        return msg
-
-    def _get_dvclose(self, dir_pth):
-        """Get outer_dvclose value from MODFLOW 6 ims file"""
-        dvclose = None
-        files = os.listdir(dir_pth)
-        for file_name in files:
-            pth = os.path.join(dir_pth, file_name)
-            if os.path.isfile(pth):
-                if file_name.lower().endswith(".ims"):
-                    with open(pth) as f:
-                        lines = f.read().splitlines()
-                    for line in lines:
-                        if "outer_dvclose" in line.lower():
-                            v = float(line.split()[1])
-                            if dvclose is None:
-                                dvclose = v
-                            else:
-                                if v > dvclose:
-                                    dvclose = v
-                            break
-
-        return dvclose
-
-    def _get_rclose(self, dir_pth):
-        """Get inner_rclose value from MODFLOW 6 ims file"""
-        rclose = None
-        files = os.listdir(dir_pth)
-        for file_name in files:
-            pth = os.path.join(dir_pth, file_name)
-            if os.path.isfile(pth):
-                if file_name.lower().endswith(".ims"):
-                    with open(pth) as f:
-                        lines = f.read().splitlines()
-                    for line in lines:
-                        if "inner_rclose" in line.lower():
-                            v = float(line.split()[1])
-                            if rclose is None:
-                                rclose = v
-                            else:
-                                if v > rclose:
-                                    rclose = v
-                            break
-
-        return rclose
-
-    def _regression_files(self, extensions):
-        if isinstance(extensions, str):
-            extensions = [extensions]
-        files = os.listdir(self.simpath)
-        files0 = []
-        files1 = []
-        for file_name in files:
-            fpth0 = os.path.join(self.simpath, file_name)
-            if os.path.isfile(fpth0):
-                for extension in extensions:
-                    if file_name.lower().endswith(extension):
-                        files0.append(fpth0)
-                        fpth1 = os.path.join(
-                            self.simpath, "mf6-regression", file_name
-                        )
-                        files1.append(fpth1)
-                        break
-        return files0, files1
-
-    def _compare_heads(self, msgall, extensions="hds"):
-        if isinstance(extensions, str):
-            extensions = [extensions]
-        success = True
-        files0, files1 = self._regression_files(extensions)
-        extension = "hds"
-        ipos = 0
-        for idx, (fpth0, fpth1) in enumerate(zip(files0, files1)):
-            outfile = os.path.splitext(os.path.basename(fpth0))[0]
-            outfile = os.path.join(
-                self.simpath, outfile + f".{extension}.cmp.out"
-            )
-            success_tst = devtools.compare_heads(
-                None,
-                None,
-                precision="double",
-                htol=self.htol,
-                text=self.extdict[extension],
-                outfile=outfile,
-                files1=fpth0,
-                files2=fpth1,
-                verbose=self.cmp_verbose,
-            )
-            msg = self.sfmt.format(
-                f"{self.extdict[extension]} comparison {ipos + 1}",
-                f"{self.name} ({os.path.basename(fpth0)})",
-            )
-            ipos += 1
-            print(msg)
-
-            if not success_tst:
-                success = False
-                msgall += msg + " ... FAILED\n"
-
-        return success, msgall
-
-    def _compare_concentrations(self, msgall, extensions="ucn"):
-        if isinstance(extensions, str):
-            extensions = [extensions]
-        success = True
-        files0, files1 = self._regression_files(extensions)
-        extension = "ucn"
-        ipos = 0
-        for idx, (fpth0, fpth1) in enumerate(zip(files0, files1)):
-            outfile = os.path.splitext(os.path.basename(fpth0))[0]
-            outfile = os.path.join(
-                self.simpath, outfile + f".{extension}.cmp.out"
-            )
-            success_tst = devtools.compare_heads(
-                None,
-                None,
-                precision="double",
-                htol=self.htol,
-                text=self.extdict[extension],
-                outfile=outfile,
-                files1=fpth0,
-                files2=fpth1,
-                verbose=self.cmp_verbose,
-            )
-            msg = self.sfmt.format(
-                f"{self.extdict[extension]} comparison {ipos + 1}",
-                f"{self.name} ({os.path.basename(fpth0)})",
-            )
-            ipos += 1
-            print(msg)
-
-            if not success_tst:
-                success = False
-                msgall += msg + " ... FAILED\n"
-
-        return success, msgall
-
-    def _compare_budgets(self, msgall, extensions="cbc"):
-        if isinstance(extensions, str):
-            extensions = [extensions]
-        success = True
-        files0, files1 = self._regression_files(extensions)
-        extension = "cbc"
-        ipos = 0
-        for idx, (fpth0, fpth1) in enumerate(zip(files0, files1)):
-            if os.stat(fpth0).st_size * os.stat(fpth0).st_size == 0:
+ignore_ext = (
+    ".hds",
+    ".hed",
+    ".bud",
+    ".cbb",
+    ".cbc",
+    ".ddn",
+    ".ucn",
+    ".glo",
+    ".lst",
+    ".list",
+    ".gwv",
+    ".mv",
+    ".out",
+)
+
+
+def model_setup(namefile, dst, remove_existing=True, extrafiles=None):
+    """Setup MODFLOW-based model files for autotests.
+
+    Parameters
+    ----------
+    namefile : str
+        MODFLOW-based model name file.
+    dst : str
+        destination path for comparison model or file(s)
+    remove_existing : bool
+        boolean indicating if an existing comparision model or file(s) should
+        be replaced (default is True)
+    extrafiles : str or list of str
+        list of extra files to include in the comparision
+
+    Returns
+    -------
+
+    """
+    # Construct src pth from namefile or lgr file
+    src = os.path.dirname(namefile)
+
+    # Create the destination folder, if required
+    create_dir = False
+    if os.path.exists(dst):
+        if remove_existing:
+            print("Removing folder " + dst)
+            shutil.rmtree(dst)
+            create_dir = True
+    else:
+        create_dir = True
+    if create_dir:
+        os.mkdir(dst)
+
+    # determine if a namefile is a lgr control file - get individual
+    # name files out of the lgr control file
+    namefiles = [namefile]
+    ext = os.path.splitext(namefile)[1]
+    if ".lgr" in ext.lower():
+        lines = [line.rstrip("\n") for line in open(namefile)]
+        for line in lines:
+            if len(line) < 1:
                 continue
-            outfile = os.path.splitext(os.path.basename(fpth0))[0]
-            outfile = os.path.join(
-                self.simpath, outfile + f".{extension}.cmp.out"
-            )
-            fcmp = open(outfile, "w")
+            if line[0] == "#":
+                continue
+            t = line.split()
+            if ".nam" in t[0].lower():
+                fpth = os.path.join(src, t[0])
+                namefiles.append(fpth)
 
-            # open the files
-            cbc0 = flopy.utils.CellBudgetFile(
-                fpth0, precision="double", verbose=self.cmp_verbose
-            )
-            cbc1 = flopy.utils.CellBudgetFile(
-                fpth1, precision="double", verbose=self.cmp_verbose
-            )
+    # Make list of files to copy
+    files2copy = []
+    for fpth in namefiles:
+        files2copy.append(os.path.basename(fpth))
+        ext = os.path.splitext(fpth)[1]
+        # copy additional files contained in the name file and
+        # associated package files
+        if ext.lower() == ".nam":
+            fname = os.path.abspath(fpth)
+            files2copy = files2copy + get_input_files(fname)
 
-            # build list of cbc data to retrieve
-            avail0 = cbc0.get_unique_record_names()
-            avail1 = cbc1.get_unique_record_names()
-            avail0 = [t.decode().strip() for t in avail0]
-            avail1 = [t.decode().strip() for t in avail1]
+    if extrafiles is not None:
+        if isinstance(extrafiles, str):
+            extrafiles = [extrafiles]
+        for fl in extrafiles:
+            files2copy.append(os.path.basename(fl))
 
-            # initialize list for storing totals for each budget term terms
-            cbc_keys0 = []
-            cbc_keys1 = []
-            for t in avail0:
-                t1 = t
-                if t not in avail1:
-                    # check if RCHA or EVTA is available and use that instead
-                    # should be able to remove this once v6.3.0 is released
-                    if t[:-1] in avail1:
-                        t1 = t[:-1]
-                    else:
-                        raise Exception(f"Could not find {t} in {fpth1}")
-                cbc_keys0.append(t)
-                cbc_keys1.append(t1)
+    # Copy the files
+    for f in files2copy:
+        srcf = os.path.join(src, f)
+        dstf = os.path.join(dst, f)
 
-            # get list of times and kstpkper
-            kk = cbc0.get_kstpkper()
-            times = cbc0.get_times()
+        # Check to see if dstf is going into a subfolder, and create that
+        # subfolder if it doesn't exist
+        sf = os.path.dirname(dstf)
+        if not os.path.isdir(sf):
+            os.makedirs(sf)
 
-            # process data
-            success_tst = True
-            for key, key1 in zip(cbc_keys0, cbc_keys1):
-                for idx, (k, t) in enumerate(zip(kk, times)):
-                    v0 = cbc0.get_data(kstpkper=k, text=key)[0]
-                    v1 = cbc1.get_data(kstpkper=k, text=key1)[0]
-                    if v0.dtype.names is not None:
-                        v0 = v0["q"]
-                        v1 = v1["q"]
-                    # skip empty vectors
-                    if v0.size < 1:
-                        continue
-                    vmin = self.rclose
-                    if vmin < 1e-6:
-                        vmin = 1e-6
-                    vmin_tol = 5.0 * vmin
-                    idx = (abs(v0) > vmin) & (abs(v1) > vmin)
-                    diff = np.zeros(v0.shape, dtype=v0.dtype)
-                    diff[idx] = abs(v0[idx] - v1[idx])
-                    diffmax = diff.max()
-                    indices = np.where(diff == diffmax)[0]
-                    if diffmax > vmin_tol:
-                        success_tst = False
-                        msg = (
-                            f"{os.path.basename(fpth0)} - "
-                            + f"{key:16s} "
-                            + f"difference ({diffmax:10.4g}) "
-                            + f"> {self.pdtol:10.4g} "
-                            + f"at {indices.size} nodes "
-                            + f" [first location ({indices[0] + 1})] "
-                            + f"at time {t} "
-                        )
-                        fcmp.write(f"{msg}\n")
-                        if self.cmp_verbose:
-                            print(msg)
+        # Now copy the file
+        if os.path.exists(srcf):
+            print("Copy file '" + srcf + "' -> '" + dstf + "'")
+            shutil.copy(srcf, dstf)
+        else:
+            print(srcf + " does not exist")
 
-            msg = self.sfmt.format(
-                f"{self.extdict[extension]} comparison {ipos + 1}",
-                f"{self.name} ({os.path.basename(fpth0)})",
-            )
-            ipos += 1
-            print(msg)
 
-            fcmp.close()
+def setup_comparison(namefile, dst, remove_existing=True):
+    """Setup a comparison model or comparision file(s) for a MODFLOW-based
+    model.
 
-            if not success_tst:
-                success = False
-                msgall += msg + " ... FAILED\n"
+    Parameters
+    ----------
+    namefile : str
+        MODFLOW-based model name file.
+    dst : str
+        destination path for comparison model or file(s)
+    remove_existing : bool
+        boolean indicating if an existing comparision model or file(s) should
+        be replaced (default is True)
 
-        return success, msgall
+
+    Returns
+    -------
+
+    """
+    # Construct src pth from namefile
+    src = os.path.dirname(namefile)
+    action = None
+    for root, dirs, files in os.walk(src):
+        dl = [d.lower() for d in dirs]
+        if any(".cmp" in s for s in dl):
+            idx = None
+            for jdx, d in enumerate(dl):
+                if ".cmp" in d:
+                    idx = jdx
+                    break
+            if idx is not None:
+                if "mf2005.cmp" in dl[idx] or "mf2005" in dl[idx]:
+                    action = dirs[idx]
+                elif "mfnwt.cmp" in dl[idx] or "mfnwt" in dl[idx]:
+                    action = dirs[idx]
+                elif "mfusg.cmp" in dl[idx] or "mfusg" in dl[idx]:
+                    action = dirs[idx]
+                elif "mf6.cmp" in dl[idx] or "mf6" in dl[idx]:
+                    action = dirs[idx]
+                elif "libmf6.cmp" in dl[idx] or "libmf6" in dl[idx]:
+                    action = dirs[idx]
+                else:
+                    action = dirs[idx]
+                break
+    if action is not None:
+        dst = os.path.join(dst, f"{action}")
+        if not os.path.isdir(dst):
+            try:
+                os.mkdir(dst)
+            except:
+                print("Could not make " + dst)
+        # clean directory
+        else:
+            print(f"cleaning...{dst}")
+            for root, dirs, files in os.walk(dst):
+                for f in files:
+                    tpth = os.path.join(root, f)
+                    print(f"  removing...{tpth}")
+                    os.remove(tpth)
+                for d in dirs:
+                    tdir = os.path.join(root, d)
+                    print(f"  removing...{tdir}")
+                    shutil.rmtree(tdir)
+        # copy files
+        cmppth = os.path.join(src, action)
+        files = os.listdir(cmppth)
+        files2copy = []
+        if action.lower() == ".cmp":
+            for file in files:
+                if ".cmp" in os.path.splitext(file)[1].lower():
+                    files2copy.append(os.path.join(cmppth, file))
+            for srcf in files2copy:
+                f = os.path.basename(srcf)
+                dstf = os.path.join(dst, f)
+                # Now copy the file
+                if os.path.exists(srcf):
+                    print("Copy file '" + srcf + "' -> '" + dstf + "'")
+                    shutil.copy(srcf, dstf)
+                else:
+                    print(srcf + " does not exist")
+        else:
+            for file in files:
+                if ".nam" in os.path.splitext(file)[1].lower():
+                    files2copy.append(
+                        os.path.join(cmppth, os.path.basename(file))
+                    )
+                    nf = os.path.join(src, action, os.path.basename(file))
+                    model_setup(nf, dst, remove_existing=remove_existing)
+                    break
+
+    return action
+
+
+def get_input_files(namefile):
+    """Return a list of all the input files in this model.
+
+    Parameters
+    ----------
+    namefile : str
+        path to a MODFLOW-based model name file
+
+    Returns
+    -------
+    filelist : list
+        list of MODFLOW-based model input files
+
+    """
+    srcdir = os.path.dirname(namefile)
+    filelist = []
+    fname = os.path.join(srcdir, namefile)
+    with open(fname, "r") as f:
+        lines = f.readlines()
+
+    for line in lines:
+        ll = line.strip().split()
+        if len(ll) < 2:
+            continue
+        if line.strip()[0] in ["#", "!"]:
+            continue
+        ext = os.path.splitext(ll[2])[1]
+        if ext.lower() not in ignore_ext:
+            if len(ll) > 3:
+                if "replace" in ll[3].lower():
+                    continue
+            filelist.append(ll[2])
+
+    # Now go through every file and look for other files to copy,
+    # such as 'OPEN/CLOSE'.  If found, then add that file to the
+    # list of files to copy.
+    otherfiles = []
+    for fname in filelist:
+        fname = os.path.join(srcdir, fname)
+        try:
+            f = open(fname, "r")
+            for line in f:
+
+                # Skip invalid lines
+                ll = line.strip().split()
+                if len(ll) < 2:
+                    continue
+                if line.strip()[0] in ["#", "!"]:
+                    continue
+
+                if "OPEN/CLOSE" in line.upper():
+                    for i, s in enumerate(ll):
+                        if "OPEN/CLOSE" in s.upper():
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            otherfiles.append(stmp)
+                            break
+        except:
+            print(fname + " does not exist")
+
+    filelist = filelist + otherfiles
+
+    return filelist
+
+
+def get_namefiles(pth, exclude=None):
+    """Search through a path (pth) for all .nam files.
+
+    Parameters
+    ----------
+    pth : str
+        path to model files
+    exclude : str or lst
+        File or list of files to exclude from the search (default is None)
+
+    Returns
+    -------
+    namefiles : lst
+        List of namefiles with paths
+
+    """
+    namefiles = []
+    for root, _, files in os.walk(pth):
+        namefiles += [
+            os.path.join(root, file) for file in files if file.endswith(".nam")
+        ]
+    if exclude is not None:
+        if isinstance(exclude, str):
+            exclude = [exclude]
+        exclude = [e.lower() for e in exclude]
+        pop_list = []
+        for namefile in namefiles:
+            for e in exclude:
+                if e in namefile.lower():
+                    pop_list.append(namefile)
+        for e in pop_list:
+            namefiles.remove(e)
+
+    return namefiles
+
+
+def get_entries_from_namefile(namefile, ftype=None, unit=None, extension=None):
+    """Get entries from a namefile. Can select using FTYPE, UNIT, or file
+    extension.
+
+    Parameters
+    ----------
+    namefile : str
+        path to a MODFLOW-based model name file
+    ftype : str
+        package type
+    unit : int
+        file unit number
+    extension : str
+        file extension
+
+    Returns
+    -------
+    entries : list of tuples
+        list of tuples containing FTYPE, UNIT, FNAME, STATUS for each
+        namefile entry that meets a user-specified value.
+
+    """
+    entries = []
+    f = open(namefile, "r")
+    for line in f:
+        if line.strip() == "":
+            continue
+        if line[0] == "#":
+            continue
+        ll = line.strip().split()
+        if len(ll) < 3:
+            continue
+        status = "UNKNOWN"
+        if len(ll) > 3:
+            status = ll[3].upper()
+        if ftype is not None:
+            if ftype.upper() == ll[0].upper():
+                filename = os.path.join(os.path.split(namefile)[0], ll[2])
+                entries.append((filename, ll[0], ll[1], status))
+        elif unit is not None:
+            if int(unit) == int(ll[1]):
+                filename = os.path.join(os.path.split(namefile)[0], ll[2])
+                entries.append((filename, ll[0], ll[1], status))
+        elif extension is not None:
+            filename = os.path.join(os.path.split(namefile)[0], ll[2])
+            ext = os.path.splitext(filename)[1]
+            if len(ext) > 0:
+                if ext[0] == ".":
+                    ext = ext[1:]
+                if extension.lower() == ext.lower():
+                    entries.append((filename, ll[0], ll[1], status))
+    f.close()
+    if len(entries) < 1:
+        entries.append((None, None, None, None))
+    return entries
+
+
+def get_sim_name(namefiles, rootpth=None):
+    """Get simulation name.
+
+    Parameters
+    ----------
+    namefiles : str or list of strings
+        path(s) to MODFLOW-based model name files
+    rootpth : str
+        optional root directory path (default is None)
+
+    Returns
+    -------
+    simname : list
+        list of namefiles without the file extension
+
+    """
+    if isinstance(namefiles, str):
+        namefiles = [namefiles]
+    sim_name = []
+    for namefile in namefiles:
+        t = namefile.split(os.sep)
+        if rootpth is None:
+            idx = -1
+        else:
+            idx = t.index(os.path.split(rootpth)[1])
+
+        # build dst with everything after the rootpth and before
+        # the namefile file name.
+        dst = ""
+        if idx < len(t):
+            for d in t[idx + 1 : -1]:
+                dst += f"{d}_"
+
+        # add namefile basename without extension
+        dst += t[-1].replace(".nam", "")
+        sim_name.append(dst)
+
+    return sim_name
+
+
+def setup_mf6(
+    src, dst, mfnamefile="mfsim.nam", extrafiles=None, remove_existing=True
+):
+    """Copy all of the MODFLOW 6 input files from the src directory to the dst
+    directory.
+
+    Parameters
+    ----------
+    src : src
+        directory path with original MODFLOW 6 input files
+    dst : str
+        directory path that original MODFLOW 6 input files will be copied to
+    mfnamefile : str
+        optional MODFLOW 6 simulation name file (default is mfsim.nam)
+    extrafiles : bool
+        boolean indicating if extra files should be included (default is None)
+    remove_existing : bool
+        boolean indicating if existing file in dst should be removed (default
+        is True)
+
+    Returns
+    -------
+    mf6inp : list
+        list of MODFLOW 6 input files
+    mf6outp : list
+        list of MODFLOW 6 output files
+
+    """
+
+    # Create the destination folder
+    create_dir = False
+    if os.path.exists(dst):
+        if remove_existing:
+            print("Removing folder " + dst)
+            shutil.rmtree(dst)
+            create_dir = True
+    else:
+        create_dir = True
+    if create_dir:
+        os.makedirs(dst)
+
+    # Make list of files to copy
+    fname = os.path.join(src, mfnamefile)
+    fname = os.path.abspath(fname)
+    mf6inp, mf6outp = get_mf6_files(fname)
+    files2copy = [mfnamefile] + mf6inp
+
+    # determine if there are any .ex files
+    exinp = []
+    for f in mf6outp:
+        ext = os.path.splitext(f)[1]
+        if ext.lower() == ".hds":
+            pth = os.path.join(src, f + ".ex")
+            if os.path.isfile(pth):
+                exinp.append(f + ".ex")
+    if len(exinp) > 0:
+        files2copy += exinp
+    if extrafiles is not None:
+        files2copy += extrafiles
+
+    # Copy the files
+    for f in files2copy:
+        srcf = os.path.join(src, f)
+        dstf = os.path.join(dst, f)
+
+        # Check to see if dstf is going into a subfolder, and create that
+        # subfolder if it doesn't exist
+        sf = os.path.dirname(dstf)
+        if not os.path.isdir(sf):
+            try:
+                os.mkdir(sf)
+            except:
+                print("Could not make " + sf)
+
+        # Now copy the file
+        if os.path.exists(srcf):
+            print("Copy file '" + srcf + "' -> '" + dstf + "'")
+            shutil.copy(srcf, dstf)
+        else:
+            print(srcf + " does not exist")
+
+    return mf6inp, mf6outp
+
+
+def get_mf6_comparison(src):
+    """Determine comparison type for MODFLOW 6 simulation.
+
+    Parameters
+    ----------
+    src : str
+        directory path to search for comparison types
+
+    Returns
+    -------
+    action : str
+        comparison type
+
+    """
+    action = None
+    # Possible comparison - the order matters
+    optcomp = (
+        "compare",
+        ".cmp",
+        "mf2005",
+        "mf2005.cmp",
+        "mfnwt",
+        "mfnwt.cmp",
+        "mfusg",
+        "mfusg.cmp",
+        "mflgr",
+        "mflgr.cmp",
+        "libmf6",
+        "libmf6.cmp",
+        "mf6",
+        "mf6.cmp",
+    )
+    # Construct src pth from namefile
+    action = None
+    for _, dirs, _ in os.walk(src):
+        dl = [d.lower() for d in dirs]
+        for oc in optcomp:
+            if any(oc in s for s in dl):
+                action = oc
+                break
+    return action
+
+
+def setup_mf6_comparison(src, dst, remove_existing=True):
+    """Setup comparision for MODFLOW 6 simulation.
+
+    Parameters
+    ----------
+    src : src
+        directory path with original MODFLOW 6 input files
+    dst : str
+        directory path that original MODFLOW 6 input files will be copied to
+    remove_existing : bool
+        boolean indicating if existing file in dst should be removed (default
+        is True)
+
+    Returns
+    -------
+    action : str
+        comparison type
+
+    """
+    # get the type of comparison to use (compare, mf2005, etc.)
+    action = get_mf6_comparison(src)
+
+    if action is not None:
+        dst = os.path.join(dst, f"{action}")
+        if not os.path.isdir(dst):
+            try:
+                os.mkdir(dst)
+            except:
+                print("Could not make " + dst)
+        # clean directory
+        else:
+            print(f"cleaning...{dst}")
+            for root, dirs, files in os.walk(dst):
+                for f in files:
+                    tpth = os.path.join(root, f)
+                    print(f"  removing...{tpth}")
+                    os.remove(tpth)
+                for d in dirs:
+                    tdir = os.path.join(root, d)
+                    print(f"  removing...{tdir}")
+                    shutil.rmtree(tdir)
+        # copy files
+        cmppth = os.path.join(src, action)
+        files = os.listdir(cmppth)
+        files2copy = []
+        if action.lower() == "compare" or action.lower() == ".cmp":
+            for file in files:
+                if ".cmp" in os.path.splitext(file)[1].lower():
+                    files2copy.append(os.path.join(cmppth, file))
+            for srcf in files2copy:
+                f = os.path.basename(srcf)
+                dstf = os.path.join(dst, f)
+                # Now copy the file
+                if os.path.exists(srcf):
+                    print("Copy file '" + srcf + "' -> '" + dstf + "'")
+                    shutil.copy(srcf, dstf)
+                else:
+                    print(srcf + " does not exist")
+        else:
+            if "mf6" in action.lower():
+                for file in files:
+                    if "mfsim.nam" in file.lower():
+                        srcf = os.path.join(cmppth, os.path.basename(file))
+                        files2copy.append(srcf)
+                        srcdir = os.path.join(src, action)
+                        setup_mf6(srcdir, dst, remove_existing=remove_existing)
+                        break
+            else:
+                for file in files:
+                    if ".nam" in os.path.splitext(file)[1].lower():
+                        srcf = os.path.join(cmppth, os.path.basename(file))
+                        files2copy.append(srcf)
+                        nf = os.path.join(src, action, os.path.basename(file))
+                        model_setup(nf, dst, remove_existing=remove_existing)
+                        break
+
+    return action
+
+
+def get_mf6_nper(tdisfile):
+    """Return the number of stress periods in the MODFLOW 6 model.
+
+    Parameters
+    ----------
+    tdisfile : str
+        path to the TDIS file
+
+    Returns
+    -------
+    nper : int
+        number of stress periods in the simulation
+
+    """
+    with open(tdisfile, "r") as f:
+        lines = f.readlines()
+    line = [line for line in lines if "NPER" in line.upper()][0]
+    nper = line.strip().split()[1]
+    return nper
+
+
+def get_mf6_mshape(disfile):
+    """Return the shape of the MODFLOW 6 model.
+
+    Parameters
+    ----------
+    disfile : str
+        path to a MODFLOW 6 discretization file
+
+    Returns
+    -------
+    mshape : tuple
+        tuple with the shape of the MODFLOW 6 model.
+
+    """
+    with open(disfile, "r") as f:
+        lines = f.readlines()
+
+    d = {}
+    for line in lines:
+
+        # Skip over blank and commented lines
+        ll = line.strip().split()
+        if len(ll) < 2:
+            continue
+        if line.strip()[0] in ["#", "!"]:
+            continue
+
+        for key in ["NODES", "NCPL", "NLAY", "NROW", "NCOL"]:
+            if ll[0].upper() in key:
+                d[key] = int(ll[1])
+
+    if "NODES" in d:
+        mshape = (d["NODES"],)
+    elif "NCPL" in d:
+        mshape = (d["NLAY"], d["NCPL"])
+    elif "NLAY" in d:
+        mshape = (d["NLAY"], d["NROW"], d["NCOL"])
+    else:
+        print(d)
+        raise Exception("Could not determine model shape")
+    return mshape
+
+
+def get_mf6_files(mfnamefile):
+    """Return a list of all the MODFLOW 6 input and output files in this model.
+
+    Parameters
+    ----------
+    mfnamefile : str
+        path to the MODFLOW 6 simulation name file
+
+    Returns
+    -------
+    filelist : list
+        list of MODFLOW 6 input files in a simulation
+    outplist : list
+        list of MODFLOW 6 output files in a simulation
+
+    """
+
+    srcdir = os.path.dirname(mfnamefile)
+    filelist = []
+    outplist = []
+
+    filekeys = ["TDIS6", "GWF6", "GWT", "GWF6-GWF6", "GWF-GWT", "IMS6"]
+    namefilekeys = ["GWF6", "GWT"]
+    namefiles = []
+
+    with open(mfnamefile) as f:
+
+        # Read line and skip comments
+        lines = f.readlines()
+
+    for line in lines:
+
+        # Skip over blank and commented lines
+        ll = line.strip().split()
+        if len(ll) < 2:
+            continue
+        if line.strip()[0] in ["#", "!"]:
+            continue
+
+        for key in filekeys:
+            if key in ll[0].upper():
+                fname = ll[1]
+                filelist.append(fname)
+
+        for key in namefilekeys:
+            if key in ll[0].upper():
+                fname = ll[1]
+                namefiles.append(fname)
+
+    # Go through name files and get files
+    for namefile in namefiles:
+        fname = os.path.join(srcdir, namefile)
+        with open(fname, "r") as f:
+            lines = f.readlines()
+        insideblock = False
+
+        for line in lines:
+            ll = line.upper().strip().split()
+            if len(ll) < 2:
+                continue
+            if ll[0] in "BEGIN" and ll[1] in "PACKAGES":
+                insideblock = True
+                continue
+            if ll[0] in "END" and ll[1] in "PACKAGES":
+                insideblock = False
+
+            if insideblock:
+                ll = line.strip().split()
+                if len(ll) < 2:
+                    continue
+                if line.strip()[0] in ["#", "!"]:
+                    continue
+                filelist.append(ll[1])
+
+    # Recursively go through every file and look for other files to copy,
+    # such as 'OPEN/CLOSE' and 'TIMESERIESFILE'.  If found, then
+    # add that file to the list of files to copy.
+    flist = filelist
+    # olist = outplist
+    while True:
+        olist = []
+        flist, olist = _get_mf6_external_files(srcdir, olist, flist)
+        # add to filelist
+        if len(flist) > 0:
+            filelist = filelist + flist
+        # add to outplist
+        if len(olist) > 0:
+            outplist = outplist + olist
+        # terminate loop if no additional files
+        # if len(flist) < 1 and len(olist) < 1:
+        if len(flist) < 1:
+            break
+
+    return filelist, outplist
+
+
+def _get_mf6_external_files(srcdir, outplist, files):
+    """Get list of external files in a MODFLOW 6 simulation.
+
+    Parameters
+    ----------
+    srcdir : str
+        path to a directory containing a MODFLOW 6 simulation
+    outplist : list
+        list of output files in a MODFLOW 6 simulation
+    files : list
+        list of MODFLOW 6 name files
+
+    Returns
+    -------
+
+    """
+    extfiles = []
+
+    for fname in files:
+        fname = os.path.join(srcdir, fname)
+        try:
+            f = open(fname, "r")
+            for line in f:
+
+                # Skip invalid lines
+                ll = line.strip().split()
+                if len(ll) < 2:
+                    continue
+                if line.strip()[0] in ["#", "!"]:
+                    continue
+
+                if "OPEN/CLOSE" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "OPEN/CLOSE":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            extfiles.append(stmp)
+                            break
+
+                if "TS6" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "FILEIN":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            extfiles.append(stmp)
+                            break
+
+                if "TAS6" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "FILEIN":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            extfiles.append(stmp)
+                            break
+
+                if "OBS6" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "FILEIN":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            extfiles.append(stmp)
+                            break
+
+                if "EXTERNAL" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "EXTERNAL":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            extfiles.append(stmp)
+                            break
+
+                if "FILE" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "FILEIN":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            extfiles.append(stmp)
+                            break
+
+                if "FILE" in line.upper():
+                    for i, s in enumerate(ll):
+                        if s.upper() == "FILEOUT":
+                            stmp = ll[i + 1]
+                            stmp = stmp.replace('"', "")
+                            stmp = stmp.replace("'", "")
+                            outplist.append(stmp)
+                            break
+
+        except:
+            print("could not get a list of external mf6 files")
+
+    return extfiles, outplist
+
+
+def get_mf6_ftypes(namefile, ftypekeys):
+    """Return a list of FTYPES that are in the name file and in ftypekeys.
+
+    Parameters
+    ----------
+    namefile : str
+        path to a MODFLOW 6 name file
+    ftypekeys : list
+        list of desired FTYPEs
+
+    Returns
+    -------
+    ftypes : list
+        list of FTYPES that match ftypekeys in namefile
+
+    """
+    with open(namefile, "r") as f:
+        lines = f.readlines()
+
+    ftypes = []
+    for line in lines:
+
+        # Skip over blank and commented lines
+        ll = line.strip().split()
+        if len(ll) < 2:
+            continue
+        if line.strip()[0] in ["#", "!"]:
+            continue
+
+        for key in ftypekeys:
+            if ll[0].upper() in key:
+                ftypes.append(ll[0])
+
+    return ftypes
+
+
+def get_mf6_blockdata(f, blockstr):
+    """Return list with all non comments between start and end of block
+    specified by blockstr.
+
+    Parameters
+    ----------
+    f : file object
+        open file object
+    blockstr : str
+        name of block to search
+
+    Returns
+    -------
+    data : list
+        list of data in specified block
+
+    """
+    data = []
+
+    # find beginning of block
+    for line in f:
+        if line[0] != "#":
+            t = line.split()
+            if t[0].lower() == "begin" and t[1].lower() == blockstr.lower():
+                break
+    for line in f:
+        if line[0] != "#":
+            t = line.split()
+            if t[0].lower() == "end" and t[1].lower() == blockstr.lower():
+                break
+            else:
+                data.append(line.rstrip())
+    return data
