--- modflow_devtools/test/conftest.py
@@ -1,101 +1,0 @@
-from os import environ
-from pathlib import Path
-
-import pytest
-from github import Github
-
-proj_root = Path(__file__).parent.parent.parent.parent
-
-
-# keepable temporary directory fixtures for various scopes
-
-
-@pytest.fixture(scope="function")
-def tmpdir(tmpdir_factory, request) -> Path:
-    node = (
-        request.node.name.replace("/", "_")
-        .replace("\\", "_")
-        .replace(":", "_")
-    )
-    temp = Path(tmpdir_factory.mktemp(node))
-    yield Path(temp)
-
-    keep = request.config.getoption("--keep")
-    if keep:
-        copytree(temp, Path(keep) / temp.name)
-
-    keep_failed = request.config.getoption("--keep-failed")
-    if keep_failed and request.node.rep_call.failed:
-        copytree(temp, Path(keep_failed) / temp.name)
-
-
-@pytest.fixture(scope="class")
-def class_tmpdir(tmpdir_factory, request) -> Path:
-    assert (
-        request.cls is not None
-    ), "Class-scoped temp dir fixture must be used on class"
-    temp = Path(tmpdir_factory.mktemp(request.cls.__name__))
-    yield temp
-
-    keep = request.config.getoption("--keep")
-    if keep:
-        copytree(temp, Path(keep) / temp.name)
-
-
-@pytest.fixture(scope="module")
-def module_tmpdir(tmpdir_factory, request) -> Path:
-    temp = Path(tmpdir_factory.mktemp(request.module.__name__))
-    yield temp
-
-    keep = request.config.getoption("--keep")
-    if keep:
-        copytree(temp, Path(keep) / temp.name)
-
-
-@pytest.fixture(scope="session")
-def session_tmpdir(tmpdir_factory, request) -> Path:
-    temp = Path(tmpdir_factory.mktemp(request.session.name))
-    yield temp
-
-    keep = request.config.getoption("--keep")
-    if keep:
-        copytree(temp, Path(keep) / temp.name)
-
-
-# misc fixtures
-
-
-@pytest.fixture
-def gh_api() -> Github:
-    return Github(environ.get("GITHUB_TOKEN"))
-
-
-@pytest.fixture
-def modflow6_path() -> Path:
-    return Path(environ.get("MODFLOW6_PATH", proj_root / "modflow6"))
-
-
-# pytest configuration hooks
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "-K",
-        "--keep",
-        action="store",
-        default=None,
-        help="Move the contents of temporary test directories to correspondingly named subdirectories at the given "
-        "location after tests complete. This option can be used to exclude test results from automatic cleanup, "
-        "e.g. for manual inspection. The provided path is created if it does not already exist. An error is "
-        "thrown if any matching files already exist.",
-    )
-
-    parser.addoption(
-        "--keep-failed",
-        action="store",
-        default=None,
-        help="Move the contents of temporary test directories to correspondingly named subdirectories at the given "
-        "location if the test case fails. This option automatically saves the outputs of failed tests in the "
-        "given location. The path is created if it doesn't already exist. An error is thrown if files with the "
-        "same names already exist in the given location.",
-    )
