--- modflow_devtools/test/test_framework.py
@@ -1,6 +1,0 @@
-def test_build():
-    pass
-
-
-def test_run():
-    pass
