--- modflow_devtools/test/test_setup.py
@@ -1,2 +1,0 @@
-def test_model_setup():
-    pass
