+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = modflow-devtools
-version = attr: modflow_devtools.config.__version__
+version = 0.0.1
 description = modflow-devtools is a Python package containing tools for MODFLOW development.
 long_description = file: README.md, LICENSE.md
 long_description_content_type = text/markdown
@@ -42,8 +42,7 @@ packages = find:
 python_requires = >=3.7
 install_requires =
     numpy
-    requests
-    flopy
+    pytest
 
 [options.extras_require]
 lint =
@@ -56,8 +55,6 @@ test =
     %(lint)s
     coverage
     flaky
-    PyGithub
-    pytest
     pytest-cov
     pytest-dotenv
     pytest-xdist
